Moved out type constraints/deduction lattice to new file.
