+++ src/ast/analysis/TypeConstraints.cpp
@@ -1,0 +1,19 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2020, The Souffle Developers. All rights reserved.
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file TypeConstraints.h
+ *
+ ***********************************************************************/
+
+#include "ast/analysis/TypeConstraints.h"
+
+namespace souffle::ast::analysis {
+
+}
